Save bad password timestamps in memory (+cap @ 30)
This is, of course, not persistent, if your battery dies. But
the file would not get cleared if unlocking with your fingerprint,
so it could grow to be very large.
This also may conflate timestamps for other users.

But as you're usually asked for a PIN/fingerprint when restarting,
or shutting down, I feel this is an OK tradeoff. Just don't think
it's good security.